Refactor and add new xcm tests (#1681)
* Refactor xcm tests

* Test for SetError, SetAPpendix, ClaimAsset

* cleanup and doc

* `registerForeignAsset` helper

* Set correct weight and fix no-error test

* prettier
