+++ tests/tests/test-xcm/test-mock-dmp-asset-transfer.ts
@@ -1,0 +1,50 @@
+import "@moonbeam-network/api-augment";
+
+import { BN } from "@polkadot/util";
+import { expect } from "chai";
+
+import { alith } from "../../util/accounts";
+import { RELAY_SOURCE_LOCATION, relayAssetMetadata } from "../../util/assets";
+import { registerForeignAsset } from "../../util/xcm";
+import { customWeb3Request } from "../../util/providers";
+import { describeDevMoonbeam } from "../../util/setup-dev-tests";
+
+// Twelve decimal places in the moonbase relay chain's token
+const RELAY_TOKEN = 1_000_000_000_000n;
+
+const palletId = "0x6D6f646c617373746d6E67720000000000000000";
+
+describeDevMoonbeam("Mock XCM - receive downward transfer", (context) => {
+  let assetId: string;
+
+  before("Should Register an asset and set unit per sec", async function () {
+    // registerForeignAsset
+    const { registeredAssetId, events, registeredAsset } = await registerForeignAsset(
+      context,
+      RELAY_SOURCE_LOCATION,
+      relayAssetMetadata
+    );
+    assetId = registeredAssetId;
+    expect(events[1].event.method.toString()).to.eq("UnitsPerSecondChanged");
+    expect(events[4].event.method.toString()).to.eq("ExtrinsicSuccess");
+    expect(registeredAsset.owner.toHex()).to.eq(palletId.toLowerCase());
+  });
+
+  it("Should receive a downward transfer of 10 DOTs to Alith", async function () {
+    // Send RPC call to inject XCM message
+    // You can provide a message, but if you don't a downward transfer is the default
+    await customWeb3Request(context.web3, "xcm_injectDownwardMessage", [[]]);
+
+    // Create a block in which the XCM will be executed
+    await context.createBlock();
+
+    // Make sure the state has ALITH's to DOT tokens
+    let alith_dot_balance = (
+      (await context.polkadotApi.query.assets.account(assetId, alith.address)) as any
+    )
+      .unwrap()
+      ["balance"].toBigInt();
+
+    expect(alith_dot_balance).to.eq(10n * RELAY_TOKEN);
+  });
+});
