+++ tests/tests/test-xcm/test-mock-dmp-queue.ts
@@ -1,86 +1,11 @@
-import "@moonbeam-network/api-augment";
-
 import { BN, u8aToHex } from "@polkadot/util";
 import { expect } from "chai";
 
-import { alith } from "../../util/accounts";
-import { RELAY_SOURCE_LOCATION } from "../../util/assets";
 import { customWeb3Request } from "../../util/providers";
 import { describeDevMoonbeam } from "../../util/setup-dev-tests";
 
 import type { XcmVersionedXcm } from "@polkadot/types/lookup";
-
-// Twelve decimal places in the moonbase relay chain's token
-const RELAY_TOKEN = 1_000_000_000_000n;
-
-const palletId = "0x6D6f646c617373746d6E67720000000000000000";
-
-const assetMetadata = {
-  name: "DOT",
-  symbol: "DOT",
-  decimals: new BN(12),
-  isFrozen: false,
-};
-
-describeDevMoonbeam("Mock XCM - receive downward transfer", (context) => {
-  let assetId: string;
-
-  before("Should Register an asset and set unit per sec", async function () {
-    // registerForeignAsset
-    const {
-      result: { events: eventsRegister },
-    } = await context.createBlock(
-      context.polkadotApi.tx.sudo.sudo(
-        context.polkadotApi.tx.assetManager.registerForeignAsset(
-          RELAY_SOURCE_LOCATION,
-          assetMetadata,
-          new BN(1),
-          true
-        )
-      )
-    );
-    // Look for assetId in events
-    assetId = eventsRegister
-      .find(({ event: { section } }) => section.toString() === "assetManager")
-      .event.data[0].toHex()
-      .replace(/,/g, "");
-
-    // setAssetUnitsPerSecond
-    const {
-      result: { events },
-    } = await context.createBlock(
-      context.polkadotApi.tx.sudo.sudo(
-        context.polkadotApi.tx.assetManager.setAssetUnitsPerSecond(RELAY_SOURCE_LOCATION, 0, 0)
-      )
-    );
-    expect(events[1].event.method.toString()).to.eq("UnitsPerSecondChanged");
-    expect(events[4].event.method.toString()).to.eq("ExtrinsicSuccess");
-
-    // check asset in storage
-    const registeredAsset = (
-      (await context.polkadotApi.query.assets.asset(assetId)) as any
-    ).unwrap();
-    expect(registeredAsset.owner.toHex()).to.eq(palletId.toLowerCase());
-  });
-
-  it("Should receive a downward transfer of 10 DOTs to Alith", async function () {
-    // Send RPC call to inject XCM message
-    // You can provide a message, but if you don't a downward transfer is the default
-    await customWeb3Request(context.web3, "xcm_injectDownwardMessage", [[]]);
-
-    // Create a block in which the XCM will be executed
-    await context.createBlock();
-
-    // Make sure the state has ALITH's to DOT tokens
-    let alith_dot_balance = (
-      (await context.polkadotApi.query.assets.account(assetId, alith.address)) as any
-    )
-      .unwrap()
-      ["balance"].toBigInt();
-
-    expect(alith_dot_balance).to.eq(10n * RELAY_TOKEN);
-  });
-});
+import { expectOk } from "../../util/expect";
 
 describeDevMoonbeam("Mock XCMP - test XCMP execution", (context) => {
   it("Should test DMP on_initialization and on_idle", async function () {
@@ -88,7 +13,6 @@ describeDevMoonbeam("Mock XCMP - test XCMP execution", (context) => {
     const balancesPalletIndex = (metadata.asLatest.toHuman().pallets as Array<any>).find(
       (pallet) => pallet.name === "Balances"
     ).index;
-    let ownParaId = (await context.polkadotApi.query.parachainInfo.parachainId()) as any;
 
     let numMsgs = 50;
     // let's target half of then being executed
@@ -152,7 +76,7 @@ describeDevMoonbeam("Mock XCMP - test XCMP execution", (context) => {
       },
     ];
 
-    let xcmMessage = {
+    const xcmMessage = {
       V2: instructions,
     };
 
@@ -172,7 +96,9 @@ describeDevMoonbeam("Mock XCMP - test XCMP execution", (context) => {
 
     // We first fund the parent sovereign account with 1000
     // we will only withdraw 1, so no problem on this
-    await context.createBlock(context.polkadotApi.tx.balances.transfer(sovereignAddress, 1000n));
+    await expectOk(
+      context.createBlock(context.polkadotApi.tx.balances.transfer(sovereignAddress, 1000n))
+    );
 
     // now we start injecting messages
     // several
@@ -188,7 +114,7 @@ describeDevMoonbeam("Mock XCMP - test XCMP execution", (context) => {
     const events = allRecords.map(({ event }) => `${event.section}.${event.method}.${event.data}`);
 
     // lets grab at which point the dmp queue was exhausted
-    let index = events.findIndex((event) => {
+    const index = events.findIndex((event) => {
       if (event.includes("dmpQueue.WeightExhausted.")) {
         return true;
       } else {
