+++ tests/util/xcm.ts
@@ -1,4 +1,4 @@
-import { u8aToHex } from "@polkadot/util";
+import { u8aToHex, BN } from "@polkadot/util";
 import { xxhashAsU8a } from "@polkadot/util-crypto";
 
 import { DevTestContext } from "./setup-dev-tests";
@@ -6,6 +6,8 @@ import {
   CumulusPalletParachainSystemRelayStateSnapshotMessagingStateSnapshot,
   XcmVersionedXcm,
 } from "@polkadot/types/lookup";
+
+import { AssetMetadata } from "./assets";
 
 // Creates and returns the tx that overrides the paraHRMP existence
 // This needs to be inserted at every block in which you are willing to test
@@ -65,4 +67,48 @@ export function mockHrmpChannelExistanceTx(
   return context.polkadotApi.tx.system.setStorage([
     [u8aToHex(overallKey), u8aToHex(stateToInsert.toU8a())],
   ]);
+}
+
+export async function registerForeignAsset(
+  context: DevTestContext,
+  asset: any,
+  metadata: AssetMetadata,
+  unitsPerSecond?: number,
+  numAssetsWeightHint?: number
+) {
+  unitsPerSecond = unitsPerSecond != null ? unitsPerSecond : 0;
+  const {
+    result: { events: eventsRegister },
+  } = await context.createBlock(
+    context.polkadotApi.tx.sudo.sudo(
+      context.polkadotApi.tx.assetManager.registerForeignAsset(asset, metadata, new BN(1), true)
+    )
+  );
+  // Look for assetId in events
+  const registeredAssetId = eventsRegister
+    .find(({ event: { section } }) => section.toString() === "assetManager")
+    .event.data[0].toHex()
+    .replace(/,/g, "");
+
+  // setAssetUnitsPerSecond
+  const {
+    result: { events },
+  } = await context.createBlock(
+    context.polkadotApi.tx.sudo.sudo(
+      context.polkadotApi.tx.assetManager.setAssetUnitsPerSecond(
+        asset,
+        unitsPerSecond,
+        numAssetsWeightHint
+      )
+    )
+  );
+  // check asset in storage
+  const registeredAsset = (
+    (await context.polkadotApi.query.assets.asset(registeredAssetId)) as any
+  ).unwrap();
+  return {
+    registeredAssetId,
+    events,
+    registeredAsset,
+  };
 }
